Moved IconPackGrid inside InputIcon
